username changes
username changes
adding controllers for code refactoring.
and validation error handeling
