+++ .env
@@ -1,2 +1,3 @@
-DC_CONNECTION=mongodb+srv://mongoadmin:mongo%40123@defaultcluster-fjzmx.mongodb.net/PortFolio_Website?retryWrites=true&w=majority
+DB_CONNECTION1=mongodb://127.0.0.1:27017/
+DB_CONNECTION=mongodb+srv://mongoadmin:mongo%40123@defaultcluster-fjzmx.mongodb.net/PortFolio_Website?retryWrites=true&w=majority
 TOKEN_SECRET=HELLOBUDDIES
